sensor/requirements.txt
wifi
RPi.GPIO
